Change the amounts donated to raw values
Figured it is unfair to show contributions
with the commission deduced as it is important
to show how much the OP has actually donated.

Also, using the stable USD to RUB currency rate
of $1=80RUB as this was the rate I used to convert
target sums in the first place.